Completed synchronization of sync32 branch code with kForth-32 v2.0.1.
Ensure that .h, .c, and .cpp files are equivalent with
their 32-bit counterparts, except for the definition of WSIZE.
Some commented source code lines may differ in the files; however,
these differences will be removed in the near future.

Revised JumpTable entries in vm64-common.s to correspond
exactly to JumpTable entries in vm32-common.s.

Revised vm64.s to:

	provide L_execute_bc
	revise definition of L_execute
	fix definition of L_2val to be consistent with 2CONSTANT

The sync32 branch executable, kforth64, performs the same
as the master branch executable with respect to the system test
code in forth-src/system-test. Three system test files do not
currently pass: coreplus.4th, divtest.4th, and regress.4th.

The sync32 branch code has a different dictionary
structure than the master branch, and is consistent with the
dictionary structure of kForth-32 version 2.x, which allows
all word-header fields to be obtained from the "xt" or from
the "nt". This was not possible in the 1.x versions of
kForth-32 or in the earlier kForth-64 implementations.

The sync32 branch also provides new words:
	EXECUTE-BC
	MY-NAME
	NAME>INTERPRET

The following words are not yet documented in the kForth
documentation:
	TRAVERSE-WORDLIST
	NAME>STRING
	NAME>INTERPRET
	MY-NAME
	EXECUTE-BC

Files Changed in sync32 branch:

	modified:   src/ForthCompiler.cpp
	modified:   src/ForthCompiler.h
	modified:   src/ForthVM.cpp
	modified:   src/ForthVM.h
	modified:   src/ForthWords.h
	modified:   src/fbc.h
	modified:   src/vm64-common.s
	modified:   src/vm64.s